feat(engine): 64 bit 列 (ValueType::Number64) の保存・書き込み・読み出し・等値の索引・oplog・sync (段階 A)
u32 の値上限を根絶する作業の土台。 列ごとの幅にする: Number64 の紐だけ cell 8B、 既存の列は u32 のまま
(v10 の DB はそのまま開ける)。 engine の値は u64 (符号なし、 u64::MAX は空の印)。

- 保存: Column の 8B atomic (Release / Acquire)、 HimoStore は幅の分岐を stored_at / value_at / store_at に
  閉じ込め、 上は u64。 u32 の列用の get_value32 / stored_slice は 64 bit 列で panic (黙って化けない)
- 索引: LockFreeCylinder の値を u64 に (sparse の key)
- layout: 紐ごとの幅 (Layout::wide、 header の型 byte から復元)、 64 bit 列の segment は 8B 幅で確保・open。
  Memory backing と pack_dir は 64 bit 列を断る (固定 4B slot)
- FILE_VERSION 11 を最初の Number64 の define で刻む (使っていない DB は v10 のまま)。 10 / 11 を読む
- oplog / wire: Tie / TieNamed の値を u64 に。 u32 に入らない値だけ TIE64 / TIE_NAMED64 (wire tag 9 / 10)。
  v11 同士でだけ通じる前提
- API: tie 系は CellValue (u32 / u64 / 負でない i32 / i64 ... リテラルもそのまま)、 列の幅に入らない値は WAL に
  載せる前に fault (ValueOutOfRange) で弾く。 get64 / get_by_id64 / query_by_id64、 pull_raw 等は CellValue、
  get / get_by_id は u32 に入る時だけ返す (切り詰めない)、 EntityValue::Num64
- 未対応は断る: live (購読 / find_by / group / sum / 並び)、 u32 の集計 (panic)
- sync: PK bind は query_by_id64、 Ref / Tag の値は u32 に入らなければ捨てる

テスト: engine tests/number64.rs 8 本 (境界値の往復・等値・AND・書き換え / 外す / 削除、 幅に入らない値の拒否、
reopen と v11、 容量の端の cell (予約 = 上限で 8B 幅が要る)、 concurrent + oplog、 recovery の replay、
未対応の経路)、 sync tests/number64_sync.rs (2 peer + wire の往復)、 oplog wide_tie_values_round_trip。
変異 11 本で全部落ちる。 workspace 1205 passed

+++ crates/enchudb-engine/tests/bridge_cursor_lost_update.rs
@@ -128,7 +128,7 @@ fn oplog_sync_bridges_every_record_under_scheduling_pressure() {
                 Some(rec) if matches!(
                     rec.op,
                     enchudb_oplog::oplog::DecodedOp::Tie { value, himo_id, .. }
-                        if value == i && himo_id == val_hid
+                        if value == i as u64 && himo_id == val_hid
                 )
             )
         });
@@ -146,7 +146,7 @@ fn oplog_sync_bridges_every_record_under_scheduling_pressure() {
                         Some(rec) if matches!(
                             rec.op,
                             enchudb_oplog::oplog::DecodedOp::Tie { value, himo_id, .. }
-                                if value == i && himo_id == val_hid
+                                if value == i as u64 && himo_id == val_hid
                         )
                     )
                 });
